Add files via upload
Push-button buzzer code to turn on the buzzer at the press of the button. This code makes use of simple PWM using analogWrite() because the buzzer I am using (ABT-414-RC) cannot use DC current hence can't make use of the digitalWrite() function to turn it on and off. Expected Result - when button is pushed the buzzer will turn on and when button is released the buzzer will turn off.